fix bugs and modify schema to add schema header
